:shower: [#5131] Some more cleanups
